ui refactor #86dt9prdc
