+++ src/js/index.js
@@ -1,28 +1,28 @@
 import { setRegistrationFormListener } from "./handlers/registration.mjs";
 import { setLoginFormListener } from "./handlers/login.mjs";
-
-//////////////////////////////////////////////////////////////
-import { themeSelector } from "./themeSelector.js";
-import { userMenuProfile, closeUserMenuProfile } from "./userMenuProfile.js";
-import { userMenuPost, closeUserMenuPost } from "./userMenuPost.js";
+import { themeSelector } from "./ui/themeSelector.js";
+import { userMenuProfile, closeUserMenuProfile } from "./ui/userMenuProfile.js";
+import { userMenuPost, closeUserMenuPost } from "./ui/userMenuPost.js";
 
 const path = location.pathname;
 
 if (path === "/") {
+	// Login page
 	setLoginFormListener();
 } else if (path === "/profile/registration/") {
 	setRegistrationFormListener();
-} else {
-	// Call themeSelector function
+} else if (path === "/profile/") {
+	// Set theme
 	themeSelector();
 
-	// Call userMenuProfile function
+	// Open/Close user menu on profile
 	userMenuProfile();
-	// Close user menu when clicking outside of it
 	closeUserMenuProfile();
+} else {
+	// Set theme
+	themeSelector();
 
-	// Call userMenuPost function
+	// Open/Close on posts
 	userMenuPost();
-	// Close user menu when clicking outside of it
 	closeUserMenuPost();
 }
